requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-dotenv
email-validator

passlib[argon2]

python-jose[cryptography]
python-multipart

pytest